examples/pytorch-webcam/requirements.txt
opencv-python
numpy
Pillow
torch
torchvision
